mass (automated) fixing of indentation & the likes)

+++ src/main/java/com/zpig333/runesofwizardry/client/render/RenderDustPlaced.java
@@ -26,7 +26,7 @@ import com.zpig333.runesofwizardry.tileentity.TileEntityDustPlaced;
  *
  */
 public class RenderDustPlaced extends TileEntitySpecialRenderer {
-	
+
 	private static final ResourceLocation dustTexture = new ResourceLocation(References.texture_path + "textures/blocks/dust_top.png");
 	/* (non-Javadoc)
 	 * @see net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer#renderTileEntityAt(net.minecraft.tileentity.TileEntity, double, double, double, float, int)
@@ -35,31 +35,31 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 	@Override
 	public void renderTileEntityAt(TileEntity tileEntity, double relativeX,
 			double relativeY, double relativeZ, float partialTicks, int blockDamageProgress) {
-		
+
 		if (!(tileEntity instanceof TileEntityDustPlaced)) return;//should not happen
 		TileEntityDustPlaced teDust = (TileEntityDustPlaced)tileEntity;
-		
+
 		// TODO external connectors
 		try{
 			//save GL state
 			GL11.glPushMatrix();
 			GL11.glPushAttrib(GL11.GL_ENABLE_BIT);
-			
+
 			//move the reference point from the player to the position of the block
 			GlStateManager.translate(relativeX, relativeY, relativeZ);
 			//grab the tesselator
 			Tessellator tesselator = Tessellator.getInstance();
 			WorldRenderer worldrenderer = tesselator.getWorldRenderer();
-			
+
 			//set texture
 			this.bindTexture(dustTexture);
-			
+
 			//setup flags
 			GL11.glDisable(GL11.GL_LIGHTING);//we want lighting (?)
 			GL11.glEnable(GL11.GL_BLEND);//we want transparency blending(?)
 			GL11.glDisable(GL11.GL_CULL_FACE);//visible from both faces (might solve being invisible)
 			GL11.glDepthMask(false);//hidden behind other objects
-			
+
 			//drawing logic here
 			//FIXME still sometimes invisible when first placed
 			//the central spots
@@ -80,18 +80,18 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 			for(int[]i:teDust.getExternalConnectors()){
 				drawExternalConnector(i[0], i[1], i[2], EnumFacing.getFront(i[3]), worldrenderer, tesselator);
 			}
-			
+
 		}finally{//restore GL stuff
 			GL11.glPopAttrib();
 			GL11.glPopMatrix();
 		}
-		
+
 	}
 	private static final double offset = 0.058;
 	final double y = 0.01;//y coordinate at which to draw the things
 	private void drawCenterVertexWithUV(int row, int col, int colorInt, WorldRenderer renderer, Tessellator tes){
 		Color color = new Color(colorInt);
-		GlStateManager.color(((float)color.getRed())/255F,((float)color.getGreen())/255F,((float)color.getBlue())/255F);
+		GlStateManager.color((color.getRed())/255F,(color.getGreen())/255F,(color.getBlue())/255F);
 		renderer.startDrawing(GL11.GL_QUADS);//our thing is quads (? the example uses triangles)
 		double rowBegin = row/(float)TileEntityDustPlaced.ROWS + offset;
 		double rowEnd = (row+1)/(float)TileEntityDustPlaced.ROWS - offset;
@@ -102,8 +102,8 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 				{colBegin,y,rowEnd, colBegin, rowEnd},
 				{colEnd,y,rowEnd, colEnd, rowEnd},
 				{colEnd,y,rowBegin, colEnd, rowBegin}
-			};
-		
+		};
+
 		for(double[] vertex:vertexTable){
 			renderer.addVertexWithUV(vertex[0],vertex[1],vertex[2],vertex[3],vertex[4]);
 		}
@@ -111,15 +111,15 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 	}
 	private final static double thin = 0.02;
 	private void drawInternalConnector(int row1, int col1, int row2, int col2, int color1, int color2, WorldRenderer renderer, Tessellator tes){
-		
+
 		double row1begin = row1/(float)TileEntityDustPlaced.ROWS + offset,
-			   row1end = (row1+1)/(float)TileEntityDustPlaced.ROWS - offset,
-			   row2begin = row2/(float)TileEntityDustPlaced.ROWS + offset,
-			   //row2end = (row2+1)/(float)TileEntityDustPlaced.ROWS - offset,
-			   col1begin = col1/(float)TileEntityDustPlaced.COLS + offset,
-			   col1end = (col1+1)/(float)TileEntityDustPlaced.COLS - offset,
-			   col2begin = col2/(float)TileEntityDustPlaced.COLS + offset;
-			   //col2end = (col2+1)/(float)TileEntityDustPlaced.COLS - offset;
+				row1end = (row1+1)/(float)TileEntityDustPlaced.ROWS - offset,
+				row2begin = row2/(float)TileEntityDustPlaced.ROWS + offset,
+				//row2end = (row2+1)/(float)TileEntityDustPlaced.ROWS - offset,
+				col1begin = col1/(float)TileEntityDustPlaced.COLS + offset,
+				col1end = (col1+1)/(float)TileEntityDustPlaced.COLS - offset,
+				col2begin = col2/(float)TileEntityDustPlaced.COLS + offset;
+		//col2end = (col2+1)/(float)TileEntityDustPlaced.COLS - offset;
 		double[][]vertexTable1=null,vertexTable2=null;
 		double middle;
 		if(row1==row2){//horizontal connector
@@ -136,7 +136,7 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 					{col2begin,y,row1end-thin,col2begin,row1end-thin},
 					{middle,y,row1end-thin,middle,row1end-thin}
 			};
-			
+
 		}else if(col1==col2){
 			middle = row1end + ((row2begin - row1end)/2);
 			vertexTable1 = new double[][]{
@@ -156,17 +156,17 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 		}
 		//safety
 		if(vertexTable1==null ||vertexTable2==null)return;
-		
+
 		Color color = new Color(color1);
-		GlStateManager.color(((float)color.getRed())/255F,((float)color.getGreen())/255F,((float)color.getBlue())/255F);
+		GlStateManager.color((color.getRed())/255F,(color.getGreen())/255F,(color.getBlue())/255F);
 		renderer.startDrawingQuads();
 		for(double vertex[]:vertexTable1){
 			renderer.addVertexWithUV(vertex[0],vertex[1],vertex[2],vertex[3],vertex[4]);
 		}
 		tes.draw();
-		
+
 		color=new Color(color2);
-		GlStateManager.color(((float)color.getRed())/255F,((float)color.getGreen())/255F,((float)color.getBlue())/255F);
+		GlStateManager.color((color.getRed())/255F,(color.getGreen())/255F,(color.getBlue())/255F);
 		renderer.startDrawingQuads();
 		for(double vertex[]:vertexTable2){
 			renderer.addVertexWithUV(vertex[0],vertex[1],vertex[2],vertex[3],vertex[4]);
@@ -213,7 +213,7 @@ public class RenderDustPlaced extends TileEntitySpecialRenderer {
 		}
 		if(vertexTable==null)return;//should not happen
 		Color colour = new Color(colorIn);
-		GlStateManager.color(((float)colour.getRed())/255F,((float)colour.getGreen())/255F,((float)colour.getBlue())/255F);
+		GlStateManager.color((colour.getRed())/255F,(colour.getGreen())/255F,(colour.getBlue())/255F);
 		renderer.startDrawingQuads();
 		for(double vertex[]:vertexTable){
 			renderer.addVertexWithUV(vertex[0],vertex[1],vertex[2],vertex[3],vertex[4]);
